add class pifpaf
